Build ina native library with Android library
Only aarch64 is supported for now. This isn't due to any technical
limitation, but because there isn't much reason to support other
architectures at the moment. They can be added later as needed.

No symbols are exported in libina.so currently. This commit just ensures
the build system builds the native library and includes libina.so in the
AAR.