Always completely fill `Patcher` read buffer
The current code is functionally correct, as according to the Rust
documentation for the `Read` trait:

"It is not an error if the returned value n is smaller than the buffer
size, even when the reader is not at the end of the stream yet."

See https://doc.rust-lang.org/std/io/trait.Read.html.

However, we don't have a good reason _not_ to completely fill the read
buffer upon each `read()` call, and filling it may significantly reduce
the number of syscalls necessary for a patch operation and thus improve
performance.

Some basic testing confirms the expected improvement, reducing the
number of syscalls by ~41% when patching dockerd 24.0.7 -> 25.0.2 (Arch
Linux, x86_64) and patch times by ~30% for the same binary. Patch times
were reduced by ~11% when patching godot 4.1.2 -> 4.1.3 (Arch Linux,
x86_64). There was no change in patch times when patching node 20.11.0
-> 20.11.1.