Prevent possibility of UB via unwinding across FFI
It is undefined behavior to unwind from Rust code across an FFI boundary
according to the Rust documentation [1]. Thus, to completely prevent
panics in Ina from causing UB on Android devices, we have four choices:

- Remove and lint for all possible panics in Ina
- Wrap all Ina code with `std::panic::catch_unwind()` in FFI wrappers
- Compile cdylib target
- Move Ina cdylib to a separate crate

The first is impractical considering the complexity of Ina's codebase,
particularly in the diffing code. The second is feasible, but
unhygienic. The third is trivial, but because Cargo doesn't support
specifying a panic strategy for a specific crate type, it requires
either 1) creating and using a custom profile when compiling the cdylib,
which could be forgotten or 2) unconditionally enabling 'panic =
"abort"' for all crate types, which is undesirable when tracing the
panic in a Rust-only application is desired. The fourth is also trivial,
but involves creating a separate crate only to enable a profile flag.

We decided on the third strategy with a custom profile as it seems to be
the least invasive to library consumers and Ina's codebase itself,
though if Cargo provided a method for specifying a panic strategy based
on the crate type, we would likely migrate to it. We may consider the
fourth strategy if the Ina C library gains enough unique functionality
to justify a separate crate.

Enabling 'panic = "abort"' additionally shaves off a few tens of
kilobytes from the final .so libraries, which is a nice bonus.

[1]: https://doc.rust-lang.org/std/panic/fn.catch_unwind.html