Run Android service as an isolated process
This change should drastically improve security of Ina against untrusted
inputs.

For some reason unbeknownst to me, passing a file descriptor over IPC as
a simple integer results in a "bad file descriptor" error when running
under an isolated process. However, file descriptors are allowed to be
used in isolated processes when passed over IPC according the base
Android SELinux policy. Apparently, passing it as a ParcelFileDescriptor
and then detaching the raw file descriptor from within the isolated
process works as expected.

Note that there isn't proper error handling for passing null for
oldFileFd. This will be added in a later commit.