Implement seccomp-bpf sandboxing
We had to remove support for i686-linux-android since seccompiler
doesn't support it. This in turn means we can no longer use ATD system
images in our Gradle managed devices. Because we are no longer
restricted to SDK 30 by the ATD system image, it seemed to make sense to
bump the SDK we test on to the latest, SDK 34.

The sandbox is initialized as soon as the patcher service starts. When
sandbox initialization fails, the service aborts itself so that we're
never running without the sandbox.

We don't currently have CI tests to verify the seccomp sandbox, but it
can be tested locally with `./gradlew allDevicesCheck`.

syscall parameters are currently not filtered. Future commits can likely
filter these to further tighten the sandbox.