Move compression to separate thread from I/O
Enabling multithreading for zstd even with only 1 thread yields a
significant diffing performance improvement somewhere in the range of a
~16-35% speed gain. Increasing the number of compression threads beyond
1 doesn't seem to improve performance much further, so the default
number of compression threads is now 1 with an option added so that the
user can change this if desired.

This change results in a slight but significant increase in max memory
usage when diffing, around 9-11% in my testing. I believe most users
will prefer the speed improvement over the memory reduction, so separate
compression threads are enabled by default.

Patch sizes were practically unaffected by this change, and if anything,
were slightly _reduced_ (by a few kilobytes at most).