Allow specifying decompression buffer size
This feature is very niche and unlikely to be used by most consumers.
However, it may be useful in specific situations where patching memory
usage needs to be capped or allocations cannot be made in the patching
process (such as when running in a sandboxed environment which prohibits
the mmap syscall).

We intend to use this feature for tightening the Android patching
sandbox later on.