Reuse a buffer for holding patch difference bytes
Previously, when reading difference bytes from the patch file to
subsequently apply to old bytes and reconstruct new data, we allocated a
new buffer for the difference bytes. This wasn't wrong per se, but
resulted in us allocating on every iteration of our patcher state
machine in PatcherState::Add. The number of allocation calls for
patching Trichrome were reported by the `memusage` command as follows:

malloc: 132
realloc: 9
calloc: 24271

After this change, the same patching operation gave the following
results:

malloc: 132
realloc: 9
calloc: 3

While reducing the number of calloc calls by 99.99% (man that's so fun
to say) sounds fantastic, it doesn't appear to have any meaningful
impact on patching speed. _Technically_ this change _slightly_ reduces
memory usage because malloc() over-allocates by miniscule amounts for
bookkeeping and it only has to do this thrice now instead of 24,000
times, but the real advantage to this change is for future
sandboxing improvements.

By removing allocations from the Patcher's read() implementation, we
take another step toward stricter sandboxing being possible. Namely, it
is now or nearly possible to forbid mmap calls in our syscall filter.

It's not immediately clear whether zstd is making allocations we don't
know about, so this sandboxing may not be possible to implement
immediately, but whatever the case, it should be much more possible now.