Implement a future-proof patch format
A patch header now consists of the following fields:

magic: 4 bytes
major version: 2 bytes
minor version: 2 bytes
data offset: varint

The major version is only incremented when the patch format changes in a
breaking way such that older versions of Ina can't read newer patches,
e.g., if we switched from zstd to brotli for the control compression
algorithm. The minor version is incremented when a patch format change
is made which older versions of Ina are still able to read, e.g., a new
field is added to the header. The data offset field represents the byte
offset from the end of the minor version field until the start of the
patch data section.

This format allows us to modify the patch format over time while still
working with older versions of Ina, which can simply skip the fields it
isn't aware of.

Closes #44