Set initial versions for crates
These all have reasonable interfaces and so are planned to be published
to crates.io once this repository is released to the public. However,
since the Android library isn't yet used in Accrescent and so may still
undergo major interface changes, we'll delay publishing it as a library
for now. We also have name squatting to worry about on crates.io,
whereas Maven Central etc. has domain verification, so we don't have to
worry about squatting for the Android library.